fix problem with 0 based precedence integers
this isn't a bug exactly, but things don't work in surprising ways
if the precedence returned in query is 0 based.

By switching to signed integers, Precedence::min() is way less than
anything anyone is likely to use and still leaves plenty of space
between 0 and 2^31 for precedents.

included a simple boolean logic grammar that demonstrates the failure